requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
appdirs==1.4.4
    # via python-escpos
argcomplete==3.6.2
    # via python-escpos
click==8.2.1
    # via uvicorn
fastapi==0.116.1
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.10
    # via anyio
importlib-resources==6.5.2
    # via python-escpos
pillow==11.3.0
    # via python-escpos
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
python-barcode==0.15.1
    # via python-escpos
python-escpos[usb]==3.1
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
pyusb==1.3.1
    # via python-escpos
pyyaml==6.0.2
    # via python-escpos
qrcode==8.2
    # via python-escpos
six==1.17.0
    # via python-escpos
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
typing-extensions==4.14.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn==0.35.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
